model-server/requirements.txt
dlib==19.24.2
easyocr==1.7.0
fastapi==0.103.0
numpy==1.24.3
opencv-python==4.8.0.76
Pillow==9.5.0
roboflow==1.1.3
ultralytics==8.0.159